refactor[#18]: 졸음 영상 묶음 다운로드 기능 리팩터링
